+++ docs/reference/changelog.rst
@@ -19,6 +19,7 @@ Changelog
 ------------------
 
 Providers
+=========
 
 - Craft Application can now build on a provider using a different architecture from the host architecture as long as the provider supports that architecture. Set the ``CRAFT_BUILD_ON`` environment variable to an architecture to use this.