+++ docs/reference/changelog.rst
@@ -21,7 +21,10 @@ Changelog
 Providers
 =========
 
-- Craft Application can now build on a provider using a different architecture from the host architecture as long as the provider supports that architecture. Set the ``CRAFT_BUILD_ON`` environment variable to an architecture to use this.
+- Craft Application can now build on a provider using a different architecture
+  from the host architecture as long as the provider supports that architecture.
+  Set the ``CRAFT_BUILD_ON`` environment variable to an architecture to use
+  this.
 
 Commands
 ========