+++ docs/reference/changelog.rst
@@ -24,7 +24,8 @@ Providers
 - Craft Application can now build on a provider using a different architecture
   from the host architecture as long as the provider supports that architecture.
   Set the ``CRAFT_BUILD_ON`` environment variable to an architecture to use
-  this.
+  this. When this is the case, in the provider instance the snap is installed from
+  the store rather than being injected from the host.
 
 Commands
 ========